pyproject.toml
[project]
name = "ad-testing-agents"
version = "0.1.0"
description = "Synthetic user simulation MCP server for testing ad offers"
requires-python = ">=3.11"
authors = [
    {name = "PersonalSystem"}
]

dependencies = [
    "anthropic>=0.50.0",
    "pydantic>=2.10.0",
    "mcp>=1.3.0",
    "streamlit>=1.45.0",
    "plotly>=6.0.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
